aws-s3: move XHRUpload code into a separate class
