+++ packages/@uppy/aws-s3/src/withXHRInlined.js
@@ -1,5 +1,5 @@
 /**
- * This plugin is currently a Big Mess™! The core reason for that is how this plugin
+ * This plugin is currently a A Big Hack™! The core reason for that is how this plugin
  * interacts with Uppy's current pipeline design. The pipeline can handle files in steps,
  * including preprocessing, uploading, and postprocessing steps. This plugin initially
  * was designed to do its work in a preprocessing step, and let XHRUpload deal with the
@@ -19,8 +19,10 @@
  * planned for Uppy v2.
  *
  * In the mean time, this plugin is stuck with a hackier approach: the necessary parts
- * of the XHRUpload implementation were copied into this plugin, and this plugin calls
- * into it immediately once it receives an upload URL.
+ * of the XHRUpload implementation were copied into this plugin, as the MiniXHRUpload
+ * class, and this plugin calls into it immediately once it receives an upload URL.
+ * This isn't as nicely modular as we'd like and requires us to maintain two copies of
+ * the XHRUpload code, but at least it's not horrifically broken :)
  */
 
 // If global `URL` constructor is available, use it
@@ -28,17 +30,10 @@ const URL_ = typeof URL === 'function' ? URL : require('url-parse')
 const { Plugin } = require('@uppy/core')
 const Translator = require('@uppy/utils/lib/Translator')
 const RateLimitedQueue = require('@uppy/utils/lib/RateLimitedQueue')
+const settle = require('@uppy/utils/lib/settle')
 const { RequestClient } = require('@uppy/companion-client')
 const qsStringify = require('qs-stringify')
-
-// For inlined chunks of XHRUpload:
-const cuid = require('cuid')
-const { Provider, Socket } = require('@uppy/companion-client')
-const emitSocketProgress = require('@uppy/utils/lib/emitSocketProgress')
-const getSocketHost = require('@uppy/utils/lib/getSocketHost')
-const settle = require('@uppy/utils/lib/settle')
-const EventTracker = require('@uppy/utils/lib/EventTracker')
-const ProgressTimeout = require('@uppy/utils/lib/ProgressTimeout')
+const MiniXHRUpload = require('./MiniXHRUpload')
 
 function resolveUrl (origin, link) {
   return new URL_(link, origin).toString()
@@ -86,27 +81,6 @@ function assertServerError (res) {
 // warning deduplication flag: see `getResponseData()` XHRUpload option definition
 let warnedSuccessActionStatus = false
 
-// See XHRUpload
-function buildResponseError (xhr, error) {
-  // No error message
-  if (!error) error = new Error('Upload error')
-  // Got an error message string
-  if (typeof error === 'string') error = new Error(error)
-  // Got something else
-  if (!(error instanceof Error)) {
-    error = Object.assign(new Error('Upload error'), { data: error })
-  }
-
-  error.request = xhr
-  return error
-}
-
-// See XHRUpload
-function setTypeInBlob (file) {
-  const dataWithUpdatedType = file.data.slice(0, file.data.size, file.meta.type)
-  return dataWithUpdatedType
-}
-
 module.exports = class AwsS3 extends Plugin {
   static VERSION = require('../package.json').version
 
@@ -136,9 +110,6 @@ module.exports = class AwsS3 extends Plugin {
     this.client = new RequestClient(uppy, opts)
     this.handleUpload = this.handleUpload.bind(this)
     this.requests = new RateLimitedQueue(this.opts.limit)
-
-    // Inlined from XHRUpload pending the pipeline rewrite:
-    this.uploaderEvents = Object.create(null)
   }
 
   setOptions (newOpts) {
@@ -240,7 +211,7 @@ module.exports = class AwsS3 extends Plugin {
 
           this.uppy.emit('preprocess-complete', this.uppy.getFile(file.id))
 
-          return this._uploadInlinedFromXHRUpload(file.id, index, numberOfFiles)
+          return this._uploader.uploadFile(file.id, index, numberOfFiles)
         })
         .catch((error) => {
           this.uppy.emit('upload-error', file, error)
@@ -248,8 +219,9 @@ module.exports = class AwsS3 extends Plugin {
     }))
   }
 
-  _getXHRUploadOptions (file) {
+  install () {
     const uppy = this.uppy
+    this.uppy.addUploader(this.handleUpload)
 
     // Get the response data from a successful XMLHttpRequest instance.
     // `content` is the S3 response as a string.
@@ -302,7 +274,7 @@ module.exports = class AwsS3 extends Plugin {
       return new Error(error)
     }
 
-    const overrides = {
+    const xhrOptions = {
       fieldName: 'file',
       responseUrlFieldName: 'location',
       timeout: this.opts.timeout,
@@ -310,326 +282,13 @@ module.exports = class AwsS3 extends Plugin {
       __queue: this.requests,
       responseType: 'text',
       getResponseData: this.opts.getResponseData || defaultGetResponseData,
-      getResponseError: defaultGetResponseError,
-      validateStatus (status, responseText, response) {
-        return status >= 200 && status < 300
-      }
-    }
-    const opts = {
-      ...this.opts,
-      ...(overrides || {}),
-      ...(file.xhrUpload || {}),
-      headers: {}
-    }
-    Object.assign(opts.headers, this.opts.headers)
-    if (overrides) {
-      Object.assign(opts.headers, overrides.headers)
-    }
-    if (file.xhrUpload) {
-      Object.assign(opts.headers, file.xhrUpload.headers)
+      getResponseError: defaultGetResponseError
     }
 
-    return opts
-  }
-
-  _uploadInlinedFromXHRUpload (id, current, total) {
-    const file = this.uppy.getFile(id)
-    if (file.error) {
-      throw new Error(file.error)
-    } else if (file.isRemote) {
-      return this._uploadRemoteInlinedFromXHRUpload(file, current, total)
-    }
-    return this._uploadLocalInlinedFromXHRUpload(file, current, total)
-  }
-
-  _addMetadataInlinedFromXHRUpload (formData, meta, opts) {
-    const metaFields = Array.isArray(opts.metaFields)
-      ? opts.metaFields
-      // Send along all fields by default.
-      : Object.keys(meta)
-    metaFields.forEach((item) => {
-      formData.append(item, meta[item])
-    })
-  }
-
-  _createFormDataUploadInlinedFromXHRUpload (file, opts) {
-    const formPost = new FormData()
-
-    this._addMetadataInlinedFromXHRUpload(formPost, file.meta, opts)
-
-    const dataWithUpdatedType = setTypeInBlob(file)
-
-    if (file.name) {
-      formPost.append(opts.fieldName, dataWithUpdatedType, file.meta.name)
-    } else {
-      formPost.append(opts.fieldName, dataWithUpdatedType)
-    }
-
-    return formPost
-  }
-
-  _createBareUploadInlinedFromXHRUpload (file, opts) {
-    return file.data
-  }
-
-  _onFileRemoveInlinedFromXHRUpload (fileID, cb) {
-    this.uploaderEvents[fileID].on('file-removed', (file) => {
-      if (fileID === file.id) cb(file.id)
-    })
-  }
-
-  _onRetryInlinedFromXHRUpload (fileID, cb) {
-    this.uploaderEvents[fileID].on('upload-retry', (targetFileID) => {
-      if (fileID === targetFileID) {
-        cb()
-      }
-    })
-  }
-
-  _onRetryAllInlinedFromXHRUpload (fileID, cb) {
-    this.uploaderEvents[fileID].on('retry-all', (filesToRetry) => {
-      if (!this.uppy.getFile(fileID)) return
-      cb()
-    })
-  }
-
-  _onCancelAllInlinedFromXHRUpload (fileID, cb) {
-    this.uploaderEvents[fileID].on('cancel-all', () => {
-      if (!this.uppy.getFile(fileID)) return
-      cb()
-    })
-  }
-
-  _uploadLocalInlinedFromXHRUpload (file, current, total) {
-    const opts = this._getXHRUploadOptions(file)
-
-    this.uppy.log(`uploading ${current} of ${total}`)
-    return new Promise((resolve, reject) => {
-      this.uppy.emit('upload-started', file)
-
-      const data = opts.formData
-        ? this._createFormDataUploadInlinedFromXHRUpload(file, opts)
-        : this._createBareUploadInlinedFromXHRUpload(file, opts)
-
-      const timer = new ProgressTimeout(opts.timeout, () => {
-        xhr.abort()
-        queuedRequest.done()
-        const error = new Error(this.i18n('timedOut', { seconds: Math.ceil(opts.timeout / 1000) }))
-        this.uppy.emit('upload-error', file, error)
-        reject(error)
-      })
-
-      const xhr = new XMLHttpRequest()
-      this.uploaderEvents[file.id] = new EventTracker(this.uppy)
-
-      const id = cuid()
-
-      xhr.upload.addEventListener('loadstart', (ev) => {
-        this.uppy.log(`[AwsS3/XHRUpload] ${id} started`)
-      })
-
-      xhr.upload.addEventListener('progress', (ev) => {
-        this.uppy.log(`[AwsS3/XHRUpload] ${id} progress: ${ev.loaded} / ${ev.total}`)
-        // Begin checking for timeouts when progress starts, instead of loading,
-        // to avoid timing out requests on browser concurrency queue
-        timer.progress()
-
-        if (ev.lengthComputable) {
-          this.uppy.emit('upload-progress', file, {
-            uploader: this,
-            bytesUploaded: ev.loaded,
-            bytesTotal: ev.total
-          })
-        }
-      })
-
-      xhr.addEventListener('load', (ev) => {
-        this.uppy.log(`[AwsS3/XHRUpload] ${id} finished`)
-        timer.done()
-        queuedRequest.done()
-        if (this.uploaderEvents[file.id]) {
-          this.uploaderEvents[file.id].remove()
-          this.uploaderEvents[file.id] = null
-        }
-
-        if (opts.validateStatus(ev.target.status, xhr.responseText, xhr)) {
-          const body = opts.getResponseData(xhr.responseText, xhr)
-          const uploadURL = body[opts.responseUrlFieldName]
-
-          const uploadResp = {
-            status: ev.target.status,
-            body,
-            uploadURL
-          }
-
-          this.uppy.emit('upload-success', file, uploadResp)
-
-          if (uploadURL) {
-            this.uppy.log(`Download ${file.name} from ${uploadURL}`)
-          }
-
-          return resolve(file)
-        } else {
-          const body = opts.getResponseData(xhr.responseText, xhr)
-          const error = buildResponseError(xhr, opts.getResponseError(xhr.responseText, xhr))
-
-          const response = {
-            status: ev.target.status,
-            body
-          }
-
-          this.uppy.emit('upload-error', file, error, response)
-          return reject(error)
-        }
-      })
-
-      xhr.addEventListener('error', (ev) => {
-        this.uppy.log(`[AwsS3/XHRUpload] ${id} errored`)
-        timer.done()
-        queuedRequest.done()
-        if (this.uploaderEvents[file.id]) {
-          this.uploaderEvents[file.id].remove()
-          this.uploaderEvents[file.id] = null
-        }
-
-        const error = buildResponseError(xhr, opts.getResponseError(xhr.responseText, xhr))
-        this.uppy.emit('upload-error', file, error)
-        return reject(error)
-      })
-
-      xhr.open(opts.method.toUpperCase(), opts.endpoint, true)
-      // IE10 does not allow setting `withCredentials` and `responseType`
-      // before `open()` is called.
-      xhr.withCredentials = opts.withCredentials
-      if (opts.responseType !== '') {
-        xhr.responseType = opts.responseType
-      }
-
-      Object.keys(opts.headers).forEach((header) => {
-        xhr.setRequestHeader(header, opts.headers[header])
-      })
-
-      const queuedRequest = this.requests.run(() => {
-        xhr.send(data)
-        return () => {
-          timer.done()
-          xhr.abort()
-        }
-      })
-
-      this._onFileRemoveInlinedFromXHRUpload(file.id, () => {
-        queuedRequest.abort()
-        reject(new Error('File removed'))
-      })
-
-      this._onCancelAllInlinedFromXHRUpload(file.id, () => {
-        queuedRequest.abort()
-        reject(new Error('Upload cancelled'))
-      })
-    })
-  }
-
-  _uploadRemoteInlinedFromXHRUpload (file, current, total) {
-    const opts = this._getXHRUploadOptions(file)
-    return new Promise((resolve, reject) => {
-      this.uppy.emit('upload-started', file)
-
-      const fields = {}
-      const metaFields = Array.isArray(opts.metaFields)
-        ? opts.metaFields
-        // Send along all fields by default.
-        : Object.keys(file.meta)
-
-      metaFields.forEach((name) => {
-        fields[name] = file.meta[name]
-      })
-
-      const Client = file.remote.providerOptions.provider ? Provider : RequestClient
-      const client = new Client(this.uppy, file.remote.providerOptions)
-      client.post(file.remote.url, {
-        ...file.remote.body,
-        endpoint: opts.endpoint,
-        size: file.data.size,
-        fieldname: opts.fieldName,
-        metadata: fields,
-        headers: opts.headers
-      }).then((res) => {
-        const token = res.token
-        const host = getSocketHost(file.remote.companionUrl)
-        const socket = new Socket({ target: `${host}/api/${token}`, autoOpen: false })
-        this.uploaderEvents[file.id] = new EventTracker(this.uppy)
-
-        this._onFileRemoveInlinedFromXHRUpload(file.id, () => {
-          socket.send('pause', {})
-          queuedRequest.abort()
-          resolve(`upload ${file.id} was removed`)
-        })
-
-        this._onCancelAllInlinedFromXHRUpload(file.id, () => {
-          socket.send('pause', {})
-          queuedRequest.abort()
-          resolve(`upload ${file.id} was canceled`)
-        })
-
-        this._onRetryInlinedFromXHRUpload(file.id, () => {
-          socket.send('pause', {})
-          socket.send('resume', {})
-        })
-
-        this._onRetryAllInlinedFromXHRUpload(file.id, () => {
-          socket.send('pause', {})
-          socket.send('resume', {})
-        })
-
-        socket.on('progress', (progressData) => emitSocketProgress(this, progressData, file))
-
-        socket.on('success', (data) => {
-          const body = opts.getResponseData(data.response.responseText, data.response)
-          const uploadURL = body[opts.responseUrlFieldName]
-
-          const uploadResp = {
-            status: data.response.status,
-            body,
-            uploadURL
-          }
-
-          this.uppy.emit('upload-success', file, uploadResp)
-          queuedRequest.done()
-          if (this.uploaderEvents[file.id]) {
-            this.uploaderEvents[file.id].remove()
-            this.uploaderEvents[file.id] = null
-          }
-          return resolve()
-        })
-
-        socket.on('error', (errData) => {
-          const resp = errData.response
-          const error = resp
-            ? opts.getResponseError(resp.responseText, resp)
-            : Object.assign(new Error(errData.error.message), { cause: errData.error })
-          this.uppy.emit('upload-error', file, error)
-          queuedRequest.done()
-          if (this.uploaderEvents[file.id]) {
-            this.uploaderEvents[file.id].remove()
-            this.uploaderEvents[file.id] = null
-          }
-          reject(error)
-        })
-
-        const queuedRequest = this.requests.run(() => {
-          socket.open()
-          if (file.isPaused) {
-            socket.send('pause', {})
-          }
-
-          return () => socket.close()
-        })
-      })
-    })
-  }
-
-  install () {
-    this.uppy.addUploader(this.handleUpload)
+    // Revert to `this.uppy.use(XHRUpload)` once the big comment block at the top of
+    // this file is solved
+    this._uploader = new MiniXHRUpload(this.uppy, xhrOptions)
+    this._uploader.i18n = this.i18n
   }
 
   uninstall () {
